+++ R/fmtr.R
@@ -114,6 +114,8 @@
 #' \item \%M = minute
 #' \item \%S = second
 #' \item \%p = AM/PM indicator
+#' \item \%q = Quarter as integer
+#' \item \%Q = Quarter as "Q?"
 #' }
 #' 
 #' See the \code{\link[base]{strptime}} function for additional codes and 
@@ -138,6 +140,7 @@
 #' fapply(t, "%d%b%Y")                # Day abbreviated month year
 #' fapply(t, "%y-%m")                 # Two digit year - month
 #' fapply(t, "%A, %B %d")             # Weekday, unabbreviated month and date
+#' fapply(t, "%Y-%Q")                 # Year and Quarter
 #' fapply(t, "%Y-%m%-%d %H:%M:%S %p") # Common timestamp format
 #' 
 #' # Examples for formatting numbers